Font size-related preference values should be CGFloats backed by doubles https://bugs.webkit.org/show_bug.cgi?id=128445
Reviewed by Filip Pizlo.

* Shared/WebPreferencesStore.h: Changed minimumFontSize, minimumLogicalFontSize,
defaultFontSize, and defaultFixedFontSize into double-value preferences.
* UIProcess/API/Cocoa/WKPreferences.h: Changed minimumFontSize into CGFloat.
* UIProcess/API/Cocoa/WKPreferences.mm:
(-[WKPreferences minimumFontSize]): Changed return type.
(-[WKPreferences setMinimumFontSize:]): Changed parameter type.
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::updatePreferences): Changed to use double getters.



Canonical link: https://commits.webkit.org/146470@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163702 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/API/Cocoa/WKPreferences.h
@@ -35,7 +35,7 @@
 WK_API_CLASS
 @interface WKPreferences : NSObject
 
-@property (nonatomic) unsigned minimumFontSize;
+@property (nonatomic) CGFloat minimumFontSize;
 
 @end
